Improvements to Asset Handling
This set of changes overhauls the helper methods used to cache templates
and map template names to full template URLs. These methods now use the
asset-pipeline API more consistently, which delivers the benefit of
working seamlessly with different asset-pipeline configurations --
asset-hosting on a CDN, or using a custom asset-prefix, for example.

Additionally, the Angular SCE delegate has been configured to trust
asset-host URLs when an asset-host has been provided. This allows
Sprangular to offload the serving of templates and other resources to a
CDN.